refactor: randomPicker를 randomRoulette으로 변경
